data-warehouse-project/requirements.txt
# Data Warehouse Project - Python Dependencies
# ============================================

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Database Connectivity
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data Generation & Faker
Faker==21.0.0

# Configuration
PyYAML==6.0.1
python-dotenv==1.0.0

# API & Web
requests==2.31.0

# Data Validation
great-expectations==0.18.8
pydantic==2.5.3

# Jupyter & Analysis
jupyter==1.0.0
notebook==7.0.6
ipython==8.18.1
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Logging & Monitoring
loguru==0.7.2

# Code Quality
black==23.12.1
flake8==6.1.0
pylint==3.0.3

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2
